Reorganize PNTerm/PNExpansion
